requirements.txt
amqp==2.5.1
autopep8==1.4.4
billiard==3.6.1.0
celery==4.3.0
Django==2.2.6
djangorestframework==3.10.3
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
pycodestyle==2.5.0
pytz==2019.2
redis==3.3.11
sqlparse==0.3.0
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
